+++ sphere.h
@@ -9,14 +9,13 @@ class Ray;
 
 struct Sphere {
     Vec3f center;
-    double radius;
+    float radius;
     Material material;
 
-    Sphere(const Vec3f &_center, const double &_radius, Material &_material) :
+    Sphere(const Vec3f &_center, const float &_radius, Material &_material) :
         center(_center), radius(_radius), material(_material) {}
 
-    bool intersect_with_ray(const Ray &ray, double &t0) const;
+    bool intersect_with_ray(const Ray &ray, float & t0) const;
 };
-
 
 #endif // SPHERE_H
